convert texture/sprites to new Codec
This is suboptimal currently. Right now the codec system is intended for Definitions which use opcode based decoding. The Sprite/Texture loaders are built off of the definition decoding which is opcode based. A implementation "OpcodedlessDecoder" was added which passes it straight to the opcode read with a opcode of -1. For both textures and sprites, this parameter should be unused in implementations of DefinitionCodec